edittest: Build edit testcase as part of main.c of X11 demo.
Addtionally: Add gcc and clang PHONY Makefile targets to be able to build
with specific compilers.
This is more in line with the other demos included in the X11 demo.
