+++ demo/x11/Makefile
@@ -1,6 +1,5 @@
 # Install
 BIN = zahnrad
-FIX = fix
 
 # Flags
 CFLAGS += -std=c89 -pedantic -O2
@@ -8,12 +7,16 @@ CFLAGS += -std=c89 -pedantic -O2
 SRC = main.c
 OBJ = $(SRC:.c=.o)
 
+# Modes
+.PHONY: gcc
+gcc: CC = gcc
+gcc: $(BIN)
+
+.PHONY: clang
+clang: CC = clang
+clang: $(BIN)
+
 $(BIN):
 	@mkdir -p bin
 	rm -f bin/$(BIN) $(OBJS)
 	$(CC) $(SRC) $(CFLAGS) -D_POSIX_C_SOURCE=200809L -o bin/$(BIN) -lX11 -lm
-
-$(FIX):
-	@mkdir -p bin
-	rm -f bin/$(FIX)
-	$(CC) editfix.c $(CFLAGS) -D_POSIX_C_SOURCE=200809L -o bin/$(FIX) -lX11 -lm
